Move netcoreapp specific stuff to netcoreapp files
